Data mining and analysis of sports and health products on Taobao
The project aims to analyze product sales patterns and user portraits by capturing commodity transaction data and user reviews of sports health products (protein powder) from the Taobao platform, and to provide references for e-commerce companies in product and user management. As consumers pay more attention to health management, the market size of health food in the fields of weight management and sports nutrition in China continues to expand. The mainstream products in this market, such as food substitutes such as nutritional powders and mixed drinks, have a market size of approximately RMB 120 billion in 2022. The high dependence on imported protein powder shows the market demand for foreign brands. As an important channel for selling these products, the Taobao platform's data reflects market trends and consumer preferences.